Multiple types of results of handling update (sendMessage, deleteMessage, etc.), multiple results of handle update

+++ src/main/java/ru/maksarts/sobescracker/dto/telegram/Update.java
@@ -10,5 +10,5 @@ public class Update implements Serializable {
     private Message message;
     private CallbackQuery callback_query;
     private Long update_id;
-    private Long chatIdFrom;
+    private Integer chatIdFrom;
 }
